requirements.txt
accelerate==1.13.0
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
aliyun-log-python-sdk==0.8.8
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
BTrees==6.3
cachetools==5.3.2
certifi==2026.2.25
cffi==2.0.0
chardet==7.3.0
charset-normalizer==3.3.2
ci-info==0.4.0
click==8.1.7
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
dashscope==1.25.15
dateparser==1.4.0
dateutils==0.6.12
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docling==2.30.0
docling-core==2.70.2
docling-ibm-models==3.13.0
docling-parse==4.7.3
docstring_parser==0.17.0
dotenv==0.9.9
duckdb==1.5.1
easyocr==1.7.2
elastic-transport==9.2.1
elasticsearch==9.3.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
fitz==0.0.1.dev2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
importlib_resources==6.5.2
iniconfig==2.3.0
ipython==8.38.0
isodate==0.7.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
json_repair==0.58.7
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
litellm==1.76.1
langchain-core==1.2.22
llama-index-core==0.12.40
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
latex2mathml==3.79.0
lazy-loader==0.5
looseversion==1.3.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.6.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
neo4j==6.1.0
networkx==3.1
nibabel==5.4.2
ninja==1.13.0
nipype==1.11.0
nltk==3.8.1
numpy==2.2.6
ollama==0.6.1
openai==2.30.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
parso==0.8.6
pathlib==1.0.1
pdfminer.six==20231228
persistent==6.5
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
portalocker==3.2.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.1
prov==2.1.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyhocon==0.3.63
pylatexenc==2.10
PyMuPDF==1.27.2.2
pyodps==0.12.2
pyparsing==3.3.2
pypdf==6.9.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==7.4.2
python-bidi==0.6.7
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==1.0.2
pytz==2026.1.post1
pyvis==0.3.2
pyxnat==1.6.4
PyYAML==6.0.3
qdrant-client==1.17.1
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
ruamel.yaml==0.19.1
safetensors==0.7.0
schedule==1.2.2
scikit-image==0.25.2
scipy==1.15.3
semchunk==3.2.5
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.3
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tifffile==2025.5.10
tokenizers==0.22.2
tomli==2.4.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.1
traitlets==5.14.3
traits==7.1.0
transaction==5.1
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.15.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.26.16
uuid_utils==0.14.1
uvicorn==0.42.0
wcwidth==0.6.0
websocket-client==1.9.0
wget==3.2
wrapt==2.1.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zc.lockfile==4.0
ZConfig==4.3
ZODB==6.2
zodbpickle==4.3
zope.deferredimport==6.1.1
zope.interface==8.2
zope.proxy==7.1
zstandard==0.25.0
